+++ NEWS.md
@@ -1,3 +1,9 @@
+# causaloptim 0.9.7
+
+- More fixes to snprintf and LEN_LABEL
+- Fixed warnings from Balke's old cpp code
+- update bibentry
+
 # causaloptim 0.9.3
 
 - Minor bug fixes in Balke's old code (sprintf -> snprintf)